feat(indexer): Add paging to auctions per address endpoint
